+++ package.json
@@ -7,7 +7,7 @@
       "@apollo/server": "^4.6.0",
       "pg": "^8.4.0",
       "reflect-metadata": "^0.1.13",
-      "typeorm": "0.3.12"
+      "typeorm": "^0.3.14"
    },
    "keywords": [],
    "author": "",
