+++ tsconfig.json
@@ -6,6 +6,7 @@
       ],
 	  "baseUrl": "./src",
       "target": "es6",
+	  "module": "commonjs",
       "moduleResolution": "node",
       "outDir": "./build",
       "emitDecoratorMetadata": true,
